requirements.txt
fastapi
uvicorn[standard]

pydantic[email]
pydantic-settings

python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1

httpx

python-multipart
aiofiles

python-dotenv

sqlalchemy
psycopg2-binary
alembic

redis[hiredis]
slowapi

pytest
pytest-asyncio
httpx
